git-svn-id: https://faracvs.cs.uni-magdeburg.de/svn/ckaestne-coloredIDE/trunk@972 16d93a9b-e91d-0410-b077-e6ebf86c48e5

+++ CIDE_FM_guidsl/src/de/ovgu/cide/fm/guidsl/XMLReaderWriter.java
@@ -152,7 +152,7 @@ public class XMLReaderWriter {
 	}
 
 	protected static RGB getRGB(long color) {
-		return new RGB(color & 255, (color >> 8) & 255, (color >> 16) & 255);
+		return new RGB((byte)color & 255,(byte) (color >> 8) & 255, (byte)(color >> 16) & 255);
 	}
 
 	protected static long getLong(RGB color) {
